# api_docs/openapi_v1_12_1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KURGIN Formula API",
    "description": "Private API wrapper around KURGIN Score Analyzer Core SDK.",
    "version": "v1.12.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health V1",
        "operationId": "health_v1_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Ready V1",
        "description": "Readiness check: import core and run a minimal one-stone analysis.",
        "operationId": "ready_v1_v1_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/analyze/stone": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Stone Endpoint",
        "operationId": "analyze_stone_endpoint_v1_analyze_stone_post",
        "parameters": [
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoneAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoneAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/batch/json": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Batch Json",
        "operationId": "analyze_batch_json_v1_analyze_batch_json_post",
        "parameters": [
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchJsonAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchJsonAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/batch/excel": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Batch Excel",
        "operationId": "analyze_batch_excel_v1_analyze_batch_excel_post",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "RU",
              "title": "Language"
            }
          },
          {
            "name": "include_records",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Records"
            }
          },
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_batch_excel_v1_analyze_batch_excel_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/export/stone/pdf": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Export Stone Pdf Endpoint",
        "operationId": "export_stone_pdf_endpoint_v1_export_stone_pdf_post",
        "parameters": [
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PdfExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/export/batch/excel": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Export Batch Excel Endpoint",
        "operationId": "export_batch_excel_endpoint_v1_export_batch_excel_post",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "RU",
              "title": "Language"
            }
          },
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_export_batch_excel_endpoint_v1_export_batch_excel_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/export/batch/package": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Export Batch Package Endpoint",
        "operationId": "export_batch_package_endpoint_v1_export_batch_package_post",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "RU",
              "title": "Language"
            }
          },
          {
            "name": "pdf_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all_ok",
              "title": "Pdf Mode"
            }
          },
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_export_batch_package_endpoint_v1_export_batch_package_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platform/stone-card": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Platform Stone Card",
        "operationId": "platform_stone_card_v1_platform_stone_card_post",
        "parameters": [
          {
            "name": "include_raw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Raw"
            }
          },
          {
            "name": "include_experimental",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Experimental"
            }
          },
          {
            "name": "x-kurgin-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kurgin-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoneAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchJsonAnalyzeRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "default": "RU"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "include_records": {
            "type": "boolean",
            "title": "Include Records",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "BatchJsonAnalyzeRequest"
      },
      "BatchJsonAnalyzeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rows_total": {
            "type": "integer",
            "title": "Rows Total"
          },
          "rows_ok": {
            "type": "integer",
            "title": "Rows Ok"
          },
          "rows_issues": {
            "type": "integer",
            "title": "Rows Issues"
          },
          "records": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Records"
          },
          "version": {
            "additionalProperties": true,
            "type": "object",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "rows_total",
          "rows_ok",
          "rows_issues",
          "version"
        ],
        "title": "BatchJsonAnalyzeResponse"
      },
      "Body_analyze_batch_excel_v1_analyze_batch_excel_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_batch_excel_v1_analyze_batch_excel_post"
      },
      "Body_export_batch_excel_endpoint_v1_export_batch_excel_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_export_batch_excel_endpoint_v1_export_batch_excel_post"
      },
      "Body_export_batch_package_endpoint_v1_export_batch_package_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_export_batch_package_endpoint_v1_export_batch_package_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PdfExportRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "default": "BILINGUAL"
          },
          "stone_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stone Result"
          }
        },
        "type": "object",
        "required": [
          "stone_result"
        ],
        "title": "PdfExportRequest"
      },
      "StoneAnalyzeRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "description": "RU or EN localization for generated text",
            "default": "RU"
          },
          "stone": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stone",
            "description": "Stone parameters using KURGIN canonical fields or aliases"
          }
        },
        "type": "object",
        "required": [
          "stone"
        ],
        "title": "StoneAnalyzeRequest"
      },
      "StoneAnalyzeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "version": {
            "additionalProperties": true,
            "type": "object",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "result",
          "version"
        ],
        "title": "StoneAnalyzeResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}